Support dependency on charging for unattended reboot
The current mechanism of scheduled unattended reboots happening at a
fixed time window are not working well for Wear, since users maybe
wearing the watch and sleeping while this reboot triggers, causing
disturbance due to audio/vibration from the reboot.

This change implements a generic feature that can be used by any device
that wishes to require charging when an unattended reboot happens. The
config is set to off by default, so that other devices remain
unaffected. We will enable the config for Wear.

Bug: 322076175
Test: unit tests
Change-Id: Icb7fdca1d1cdc25b0854d09b8de5e6dfc03b99bf