Fix ConfigInfra Resource Package Fetching
The resources package name maybe changed during build-time, so the
current approach that looks for the default package name wouldn't work
in those cases.

This change checks for a package-name suffix, instead of trying to look
for an exact package-name (among system packages).

After this change is merged, we will send changes to use the new
ServiceResourcesHelper in other places within the server code.

Bug: 324961721
Test: unit tests

Change-Id: Iea964270a1ed3f34ddc20cb586b70c12e37e5b43